+++ src/components/content/forms/patients.ts
@@ -1,54 +1,59 @@
-import {
-  PatientHistoryResponseError,
-  schedulePatientHistory,
-} from "@/api/patient-history";
 import { CTWRequestContext } from "@/components/core/ctw-context";
 import { createOrEditFhirResource } from "@/fhir/action-helper";
 import { dateToISO } from "@/fhir/formatters";
 import { PatientModel } from "@/fhir/models";
-import { getFormStatusErrors } from "@/utils/errors";
-import { AnyZodSchema, getFormData } from "@/utils/form-helper";
 import { QUERY_KEY_PATIENT } from "@/utils/query-keys";
 import { queryClient } from "@/utils/request";
 
+export type PatientFormData = {
+  lastName: string;
+  firstName: string;
+  gender: fhir4.Patient["gender"];
+  dateOfBirth: Date;
+  email: string;
+  phone: string;
+  address: string;
+  city: string;
+  state: string;
+  zipCode: string;
+};
+
 export const editPatient = async (
   patient: PatientModel,
-  data: FormData,
-  getRequestContext: () => Promise<CTWRequestContext>,
-  schema: AnyZodSchema
+  data: PatientFormData,
+  getRequestContext: () => Promise<CTWRequestContext>
 ) => {
-  const result = await getFormData(data, schema);
-  let requestErrors: string[] = [];
-
-  if (!result.success) {
-    return { formResult: result, requestErrors: undefined };
-  }
-
   const requestContext = await getRequestContext();
 
   const fhirPatient: fhir4.Patient = {
     resourceType: "Patient",
     id: patient.id,
     active: patient.active,
-    name: [{ family: result.data.lastName, given: [result.data.firstName] }],
-    gender: result.data.gender,
-    birthDate: dateToISO(result.data.dateOfBirth),
+    name: [
+      {
+        family: data.lastName,
+        given: [data.firstName],
+        use: patient.use ?? "official",
+      },
+    ],
+    gender: data.gender,
+    birthDate: dateToISO(data.dateOfBirth),
     telecom: [
       {
         system: "email",
-        value: result.data.email,
+        value: data.email,
       },
       {
         system: "phone",
-        value: result.data.phone,
+        value: data.phone,
       },
     ],
     address: [
       {
-        line: [result.data.address],
-        city: result.data.city,
-        state: result.data.state,
-        postalCode: result.data.zipCode,
+        line: [data.address],
+        city: data.city,
+        state: data.state,
+        postalCode: data.zipCode,
       },
     ],
     contact: patient.contact,
@@ -60,48 +65,8 @@ export const editPatient = async (
   };
 
   const response = await createOrEditFhirResource(fhirPatient, requestContext);
-  const statusErrors = getFormStatusErrors(requestErrors, result, response);
-  requestErrors = statusErrors.errors;
-  result.success = statusErrors.resultIsSuccess;
 
   await queryClient.invalidateQueries([QUERY_KEY_PATIENT]);
 
-  return { formResult: result, requestErrors };
-};
-
-export const editPatientAndScheduleHistory = async (
-  patient: PatientModel,
-  data: FormData,
-  getRequestContext: () => Promise<CTWRequestContext>,
-  schema: AnyZodSchema
-) => {
-  const editPatientResponse = await editPatient(
-    patient,
-    data,
-    getRequestContext,
-    schema
-  );
-
-  const result = editPatientResponse.formResult;
-  let { requestErrors } = editPatientResponse;
-
-  const requestContext = await getRequestContext();
-
-  if (result.success) {
-    const patientHistoryResponse = await schedulePatientHistory(
-      requestContext,
-      patient.id,
-      result.data
-    );
-    if ("errors" in patientHistoryResponse) {
-      requestErrors = [
-        patientHistoryResponse.errors.map(
-          (err: PatientHistoryResponseError) => err.details
-        ),
-      ];
-      result.success = false;
-    }
-  }
-
-  return { formResult: result, requestErrors };
+  return response;
 };