+++ src/fhir/conditions.ts
@@ -1,7 +1,7 @@
 import { SearchParams } from "fhir-kit-client";
 import { orderBy } from "lodash";
+import { getIncludedBasics } from "./bundle";
 import { CodePreference } from "./codeable-concept";
-import { getPractitioner } from "./practitioner";
 import {
   searchBuilderRecords,
   searchCommonRecords,
@@ -16,7 +16,6 @@ import {
 } from "./system-urls";
 import { getZusApiBaseUrl } from "@/api/urls";
 import { getAddConditionWithDefaults } from "@/components/content/forms/conditions";
-import { CTWRequestContext } from "@/components/core/ctw-context";
 import { useQueryWithPatient } from "@/components/core/patient-provider";
 import { ConditionModel } from "@/fhir/models/condition";
 import { errorResponse } from "@/utils/errors";
@@ -25,6 +24,8 @@ import {
   QUERY_KEY_OTHER_PROVIDER_CONDITIONS,
   QUERY_KEY_PATIENT_CONDITIONS,
 } from "@/utils/query-keys";
+import { CTWRequestContext } from "@/components/core/ctw-context";
+import { getPractitioner } from "./practitioner";
 
 export const CONDITION_CODE_PREFERENCE_ORDER: CodePreference[] = [
   { system: SYSTEM_SNOMED, checkForEnrichment: true },
@@ -53,14 +54,14 @@ export function usePatientConditions() {
     [],
     async (requestContext, patient) => {
       try {
-        const { resources: conditions } = await searchBuilderRecords(
+        const { bundle, resources: conditions } = await searchBuilderRecords(
           "Condition",
           requestContext,
           {
             patientUPID: patient.UPID as string,
           }
         );
-        return filterAndSort(conditions);
+        return filterAndSort(setupConditionModels(conditions, bundle));
       } catch (e) {
         throw new Error(
           `Failed fetching condition information for patient: ${e}`
@@ -76,14 +77,15 @@ export function useOtherProviderConditions() {
     [],
     async (requestContext, patient) => {
       try {
-        const { resources: conditions } = await searchSummaryRecords(
+        const { bundle, resources: conditions } = await searchSummaryRecords(
           "Condition",
           requestContext,
           {
+            _revinclude: "Basic:subject",
             patientUPID: patient.UPID as string,
           }
         );
-        return filterAndSort(conditions);
+        return filterAndSort(setupConditionModels(conditions, bundle));
       } catch (e) {
         throw new Error(
           `Failed fetching condition information for patient: ${e}`
@@ -138,12 +140,24 @@ export function useConditionHistory(condition?: ConditionModel) {
   );
 }
 
-function filterAndSort(conditions: fhir4.Condition[]) {
+function setupConditionModels(
+  conditionResources: fhir4.Condition[],
+  bundle: fhir4.Bundle
+): ConditionModel[] {
+  const basicsMap = getIncludedBasics(bundle);
+  return conditionResources.map(
+    (c) => new ConditionModel(c, undefined, basicsMap[c.id ?? ""])
+  );
+}
+
+function filterAndSort(conditions: ConditionModel[]): ConditionModel[] {
   return orderBy(
-    conditions.filter((condition) => condition.asserter?.type !== "Patient"),
+    conditions.filter(
+      (condition) => condition.resource.asserter?.type !== "Patient"
+    ),
     [
-      (condition) => new ConditionModel(condition).resource.recordedDate ?? "",
-      (condition) => new ConditionModel(condition).display,
+      (condition) => condition.resource.recordedDate ?? "",
+      (condition) => condition.display,
     ],
     ["desc"]
   );